refactor: slim lib.rs by extracting router and CORS middleware
- Extract route table to api/router.rs with its own DEFERRED_ANALYTICS thread-local.
- Extract CORS/security header helpers to middleware/cors.rs.
