+++ src/api/mod.rs
@@ -6,6 +6,7 @@ pub mod keys;
 pub mod links;
 pub mod orgs;
 pub mod reports;
+pub mod router;
 pub mod settings;
 pub mod tags;
 pub mod title_fetch;
