+++ src/middleware/mod.rs
@@ -1,3 +1,5 @@
+pub mod cors;
 pub mod rate_limit;
 
+pub use cors::{add_cors_headers, add_security_headers};
 pub use rate_limit::{RateLimitConfig, RateLimiter, is_kv_rate_limiting_enabled};
